luna_v.3.3.3/luna-faces/requirements.txt
requests==2.18.4
tornado==5.0.1
logbook==1.1.0
ujson==1.35
python-interface==1.4.0
faker==0.7.10
sqlalchemy-utils==0.32.13
SQLAlchemy-migrate==0.11.0
enum34==1.1.6
jsonschema==2.6.0
SQLAlchemy==1.2.5
psycopg2==2.7.1
cx_oracle==6.3.1
pytz==2018.4
six==1.11.0
python-dateutil==2.7.3
alembic==1.0.0
extras/luna3-v.0.0.52.tar.gz
